# 08-rest/openapi.yaml
openapi: "3.0.2"
info:
  title: Thread API
  version: "1.0"
servers:
  - url: http://localhost:8080/nswi145rest/rest/

paths:
  /thread:
    get:
      description: list all threads
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Thread"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Thread"
    post:
      description: create new thread
      requestBody:
        content:
          application/xml:
            schema:
              $ref: "#/components/schemas/Thread"
        required: true
      responses:
        "201":
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                description: link to the created resource
        "409":
          description: Thread already exists
  /thread/{threadId}:
    get:
      description: return a single thread
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
            application/xml:
              schema:
                $ref: "#/components/schemas/Thread"
        "400":
          description: invalid id supplied
        "404":
          description: thread not found
    delete:
      description: delete a single thread
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: successful operation
        "400":
          description: invalid id supplied
        "404":
          description: thread not found

components:
  schemas:
    Thread:
      type: object
      properties:
        id:
          type: integer
        owner:
          type: integer
        name:
          type: string