# 08-rest/src/main/java/nswi145rest/Thread.java
package nswi145rest;

import jakarta.xml.bind.annotation.XmlElement;
import jakarta.xml.bind.annotation.XmlRootElement;

@XmlRootElement
public class Thread {
	private int id;
	private String name;
	private int owner;
	
	public Thread() {}
	public Thread(int id, String name, int owner) {
		this.id = id;
		this.name = name;
		this.owner = owner;
	}

	@XmlElement(name="id")
	public int getId() {
		return id;
	}
	public void setId(int id) {
		this.id = id;
	}
	
	@XmlElement(name="name")
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	
	@XmlElement(name="owner")
	public int getOwner() {
		return owner;
	}
	public void setOwner(int owner) {
		this.owner = owner;
	}
}